+++ src/Event.c
@@ -29,8 +29,8 @@ void Event_SetTerm(Event *event, Term term)
     event->term = term;
 }
 
-long base = 1;
-Event Event_InputEvent(Term term, char type, Truth truth, long currentTime)
+uint32_t base = 1;
+Event Event_InputEvent(Term term, char type, Truth truth, uint32_t currentTime)
 {
     return (Event) { .term = term,
                      .type = type, 
@@ -52,7 +52,7 @@ void Event_Print(Event *event)
     printf(event->type == EVENT_TYPE_GOAL ? "type=goal\n" : (EVENT_TYPE_BELIEF ? "type=belief\n" : "type=deleted\n" ));
     Truth_Print(&event->truth);
     Stamp_print(&event->stamp);
-    printf("occurrenceTime=%ld\n\n", event->occurrenceTime);
+    printf("occurrenceTime=%" PRIu32 "\n\n", event->occurrenceTime);
 }
 
 bool Event_Equal(Event *event, Event *existing)
