+++ src/Inference.h
@@ -45,7 +45,7 @@
 //Methods//
 //-------//
 //{Event a.} |- Event a. Truth_Projection (projecting to current time)
-Event Inference_EventUpdate(Event *ev, long currentTime);
+Event Inference_EventUpdate(Event *ev, uint32_t currentTime);
 //{Event a., Event b.} |- Event (&/,a,b). Truth_Intersection (after projecting b to a)
 Event Inference_BeliefIntersection(Event *a, Event *b, bool *success);
 //{Event a., Event b.} |- Implication <a =/> c>. Truth_Eternalize(Truth_Induction) (after projecting b to a)
@@ -55,9 +55,9 @@ Implication Inference_ImplicationRevision(Implication *a, Implication *b);
 //{Event b!, Implication <a =/> b>.} |- Event a! Truth_Deduction
 Event Inference_GoalDeduction(Event *component, Implication *compound);
 //{Event (&/,a,op())!, Event a.} |- Event op()! Truth_Deduction
-Event Inference_OperationDeduction(Event *compound, Event *component, long currentTime);
+Event Inference_OperationDeduction(Event *compound, Event *component, uint32_t currentTime);
 //{Event a!, Event a!} |- Event a! Truth_Revision or Choice (dependent on evidental overlap)
-Event Inference_RevisionAndChoice(Event *existing_potential, Event *incoming_spike, long currentTime, bool *revised);
+Event Inference_RevisionAndChoice(Event *existing_potential, Event *incoming_spike, uint32_t currentTime, bool *revised);
 //{Event a., Implication <a =/> b>.} |- Event b.  Truth_Deduction
 Event Inference_BeliefDeduction(Event *component, Implication *compound);
