+++ src/unit_tests/HashTable_Test.h
@@ -80,7 +80,7 @@ void HashTable_Test()
     char blockname[ATOMIC_TERM_LEN_MAX] = {0};
     strncpy(blockname, testname, ATOMIC_TERM_LEN_MAX-1);
     HashTable_Set(&HTtest2, blockname, (void*) 42);
-    long res = (long) HashTable_Get(&HTtest2, blockname);
+    int res = (int) HashTable_Get(&HTtest2, blockname);
     assert(res == 42, "Result is not right!");
     puts(">>HashTable test successul");
 }
